requirements.txt
absl-py==0.10.0
alabaster==0.7.12
allennlp==0.8.4
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.3
asn1crypto==0.24.0
astroid==2.2.5
astropy==3.2.1
astunparse==1.6.3
atomicwrites==1.3.0
attrs==20.1.0
awscli==1.18.132
Babel==2.8.0
backcall==0.1.0
backports.functools-lru-cache==1.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.7.1
bitarray==0.9.3
bkcharts==0.2
bleach==3.1.0
blis==0.2.4
bokeh==1.2.0
boto==2.49.0
boto3==1.14.55
botocore==1.17.55
Bottleneck==1.2.1
cachetools==4.1.1
certifi==2021.5.30
cffi==1.12.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.2.1
clyent==1.2.2
colorama==0.4.3
conda==4.7.10
conda-build==3.18.8
conda-package-handling==1.3.11
conda-verify==3.4.2
conllu==0.11
contextlib2==0.5.5
cryptography==2.7
cycler==0.10.0
Cython==0.29.12
cytoolz==0.10.0
dask==2.1.0
decorator==4.4.0
defusedxml==0.6.0
detectron2==0.2.1+cu101
distributed==2.1.0
docutils==0.15.2
editdistance==0.5.3
entrypoints==0.3
et-xmlfile==1.0.1
fastapi==0.63.0
fastcache==1.1.0
filelock==3.0.12
flaky==3.7.0
Flask==1.1.2
Flask-Cors==3.0.9
ftfy==5.8
future==0.17.1
fvcore==0.1.1.post20200716
gast==0.3.3
gdown==3.13.0
gevent==20.6.2
glob2==0.7
gmpy2==2.0.8
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.16
grpcio==1.31.0
gunicorn==20.0.4
h11==0.12.0
h5py==2.10.0
heapdict==1.0.0
html5lib==1.0.1
idna==2.10
imageio==2.5.0
imagesize==1.2.0
importlib-metadata==1.7.0
iniconfig==1.0.1
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.13.3
jeepney==0.4
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
json5==0.8.4
jsonnet==0.16.0
jsonpickle==1.4.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.0.2
jupyterlab-server==1.0.0
Keras-Preprocessing==1.1.2
keyring==18.0.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.1
libarchive-c==2.8
lief==0.9.0
llvmlite==0.29.0
locket==0.2.0
lxml==4.3.4
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.12
mkl-random==1.0.2
mkl-service==2.0.2
mock==3.0.5
more-itertools==8.5.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
nltk==3.5
nose==1.3.7
notebook==6.0.0
numba==0.44.1
numexpr==2.6.9
numpy==1.18.5
numpydoc==1.1.0
oauthlib==3.1.0
olefile==0.46
openpyxl==2.6.2
opt-einsum==3.3.0
overrides==3.1.0
packaging==20.4
pandas==0.24.2
pandocfilters==1.4.2
parsimonious==0.8.1
parso==0.5.0
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.4
patsy==0.5.1
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.2.0
pkginfo==1.5.0.1
plac==0.9.6
pluggy==0.13.1
ply==3.11
portalocker==2.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.13.0
psutil==5.6.3
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.1
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
pydantic==1.7.3
pydot==1.4.1
pyflakes==2.1.1
Pygments==2.6.1
pylint==2.3.1
pyodbc==4.0.26
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.14.11
PySocks==1.7.0
pytesseract==0.3.8
pytest==6.0.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytorch-pretrained-bert==0.6.2
pytorch3d==0.3.0
pytz==2020.1
PyWavelets==1.0.3
PyYAML==5.3.1
pyzmq==18.0.0
QtAwesome==0.5.7
qtconsole==4.5.1
QtPy==1.8.0
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
responses==0.12.0
rope==0.14.0
rsa==4.5
ruamel-yaml==0.15.46
s3transfer==0.3.3
sacremoses==0.0.43
scikit-image==0.15.0
scikit-learn==0.20.0
scipy==1.4.1
seaborn==0.9.0
SecretStorage==3.1.1
Send2Trash==1.5.0
sentencepiece==0.1.91
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.8
spacy==2.1.9
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.1
SQLAlchemy==1.3.5
sqlparse==0.3.1
srsly==1.0.2
starlette==0.13.6
statsmodels==0.10.0
sympy==1.4
tables==3.5.2
tabulate==0.8.7
tblib==1.4.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
termcolor==1.1.0
terminado==0.8.2
tesseract-ocr==0.0.1
testpath==0.4.2
thinc==7.0.8
toolz==0.10.0
tornado==6.0.3
tqdm==4.48.2
traitlets==4.3.2
transformers==2.2.2
typing-extensions==3.7.4.3
unicodecsv==0.14.1
Unidecode==1.1.1
urllib3==1.25.10
uvicorn==0.13.3
wasabi==0.8.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.0
word2number==1.1
wrapt==1.12.1
wurlitzer==1.0.2
xlrd==1.2.0
XlsxWriter==1.1.8
xlwt==1.3.0
yacs==0.1.8
zict==1.0.0
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
